requirements.txt
python_version = 3.12
tabulate
mysql-connector-python
pandas
matplotlib